Create iterstack.h
The class std::stack does not provide an iterator, which is needed for this class.  So, like std::stack, class iterstack is implemented on top of a container.

We use private inheritance because we want to restrict operations only to those few that are approved.  All functions are merely inherited from the container, with only ones needed being exported as public.
No implementation file is needed because all functions are inherited, and the convenience functions that are added are trivial, and so can be inline.

Any underlying container which supports the necessary operations
could be used, such as vector, list, or deque.